add relationships, reduce the data in *.sqls
It's incrediblely easy to add the relaionship to so many tables, because
there is no complicated relation, only one to many. Maybe later I need
to add 'backref' or someting like that.
the initial *.sqls is stored in other places, it's slow to load so many
data, but after I cut some, some data with foreign key many fail to
load.